Add unit tests for pacing.
