+++ UnitTest1/unittest1.cpp
@@ -438,6 +438,13 @@ namespace UnitTest1
             Assert::AreEqual(ret, 0);
         }
 
+        TEST_METHOD(pacing_repeat)
+        {
+            int ret = pacing_repeat_test();
+
+            Assert::AreEqual(ret, 0);
+        }
+
 		TEST_METHOD(test_tls_api)
 		{
 			int ret = tls_api_test();
