+++ picoquic/picoquic_internal.h
@@ -991,23 +991,24 @@ typedef struct st_picoquic_remote_cnxid_stash_t {
 * Pacing uses a set of per path variables:
 * - rate: bytes per second.
 * - evaluation_time: last time the path was evaluated.
-* - bucket_nanosec: number of nanoseconds of transmission time that are allowed.
 * - bucket_max: maximum value (capacity) of the leaky bucket.
-* - packet_time_nanosec: number of nanoseconds required to send a full size packet.
 * - packet_time_microsec: max of (packet_time_nano_sec/1024, 1) microsec.
+* Internal variables:
+* - bucket_nanosec: number of nanoseconds of transmission time that are allowed.
+* - packet_time_nanosec: number of nanoseconds required to send a full size packet.
 */
 typedef struct st_picoquic_pacing_t {
     uint64_t rate;
     uint64_t evaluation_time;
-    int64_t bucket_nanosec;
     int64_t bucket_max;
-    int64_t packet_time_nanosec;
     uint64_t packet_time_microsec;
     uint64_t quantum_max;
     uint64_t rate_max;
     int bandwidth_pause;
+    /* High precision variables should only be used inside pacing.c */
+    int64_t bucket_nanosec;
+    int64_t packet_time_nanosec;
 } picoquic_pacing_t;
-
 
 /*
 * Per path context.
@@ -1177,30 +1178,7 @@ typedef struct st_picoquic_path_t {
     uint64_t last_cwin_blocked_time;
     uint64_t last_time_acked_data_frame_sent;
     void* congestion_alg_state;
-
-#if 1
     picoquic_pacing_t pacing;
-#else
-    /*
-    * Pacing uses a set of per path variables:
-    * - pacing_rate: bytes per second.
-    * - pacing_evaluation_time: last time the path was evaluated.
-    * - pacing_bucket_nanosec: number of nanoseconds of transmission time that are allowed.
-    * - pacing_bucket_max: maximum value (capacity) of the leaky bucket.
-    * - pacing_packet_time_nanosec: number of nanoseconds required to send a full size packet.
-    * - pacing_packet_time_microsec: max of (packet_time_nano_sec/1024, 1) microsec.
-    */
-
-    uint64_t pacing_rate;
-    uint64_t pacing_evaluation_time;
-    int64_t pacing_bucket_nanosec;
-    int64_t pacing_bucket_max;
-    int64_t pacing_packet_time_nanosec;
-    uint64_t pacing_packet_time_microsec;
-    uint64_t pacing_quantum_max;
-    uint64_t pacing_rate_max;
-    int pacing_bandwidth_pause;
-#endif
 
     /* MTU safety tracking */
     uint64_t nb_mtu_losses;
@@ -1639,6 +1617,8 @@ picoquic_cnx_t* picoquic_cnx_by_icid(picoquic_quic_t* quic, picoquic_connection_
 picoquic_cnx_t* picoquic_cnx_by_secret(picoquic_quic_t* quic, const uint8_t* reset_secret, const struct sockaddr* addr);
 
 /* Pacing implementation */
+void picoquic_pacing_init(picoquic_pacing_t* pacing, uint64_t current_time);
+int picoquic_is_pacing_blocked(picoquic_pacing_t* pacing);
 int picoquic_is_authorized_by_pacing(picoquic_pacing_t* pacing, uint64_t current_time, uint64_t* next_time, unsigned int packet_train_mode, picoquic_quic_t * quic);
 void picoquic_update_pacing_parameters(picoquic_pacing_t* pacing, double pacing_rate, uint64_t quantum, size_t send_mtu, uint64_t smoothed_rtt,
     picoquic_path_t* signalled_path);
