+++ picoquictest/picoquictest.vcxproj.filters
@@ -165,6 +165,9 @@
     <ClCompile Include="openssl_test.c">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="pacing_test.c">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="picoquictest.h">
